Support max_phase quasi-enum and mapping statuses
Add support for ChEMBL max_phase as a quasi-enum (includes -1 and non-integer 0.5) and introduce mapping-status enum family (mapped/unmapped/missing). Wire the new values through configs, enum registry paths, normalization profiles and normalizers (normalize_profile_quasi_enum_numeric and normalize_profile_mapping_status), domain entities and validation (allow int|float and validate against MAX_PHASE_VALUES), and update fixtures/tests/docs to reflect the new policy. Also includes assorted cleanups and small refactors: helper extraction in cleanup_repository and neo4j script, minor formatting/typing adjustments, and removal of an extra trailing newline in a diagnostics summary file.

+++ tests/unit/infrastructure/schemas/test_silver_pipeline_contracts.py
@@ -267,7 +267,7 @@ PIPELINE_SCHEMA_EXPECTATIONS: dict[str, list[tuple[str, pa.DataType]]] = {
         ("hierarchy_parent_chembl_id", pa.string()),
         ("inchi_key", pa.string()),
         ("inorganic_flag", pa.int64()),
-        ("max_phase", pa.int64()),
+        ("max_phase", pa.float64()),
         ("molecule_id", pa.string()),
         ("molecule_hierarchy", pa.string()),
         ("molecule_properties", pa.string()),
